Issue microsoft #113: Implement an option to allow DiskSpd to output higher precision results.
Issue microsoft#114: Implement a more complete "driver" for the ResultParser(s)

. Some local re-factoring to implement microsoft#113.  This is not a complete implementation, but a first step.

+++ IORequestGenerator/IORequestGenerator.cpp
@@ -2323,7 +2323,7 @@ bool IORequestGenerator::_GenerateRequestsForTimeSpan(const Profile& profile, co
         }
 
         // each thread has a different random seed
-        Random *pRand = new Random(timeSpan.GetRandSeed() + iThread);
+        Random* pRand = new Random(static_cast<UINT64>(timeSpan.GetRandSeed()) + iThread);
         if (nullptr == pRand)
         {
             PrintError("FATAL ERROR: could not allocate memory\n");
